+++ cmd/consume/consume_test.go
@@ -1,6 +1,7 @@
 package consume_test
 
 import (
+	"fmt"
 	"github.com/deviceinsight/kafkactl/test_util"
 	"strings"
 	"testing"
@@ -123,4 +124,57 @@ func TestConsumeAutoCompletionIntegration(t *testing.T) {
 	test_util.AssertContains(t, topicName1, outputLines)
 	test_util.AssertContains(t, topicName2, outputLines)
 	test_util.AssertContains(t, topicName3, outputLines)
+}
+
+func TestAvroDeserializationErrorHandlingIntegration(t *testing.T) {
+
+	test_util.StartIntegrationTest(t)
+
+	valueSchema := `{
+  "name": "person",
+  "type": "record",
+  "fields": [
+	{
+      "name": "name",
+      "type": "string"
+    }
+  ]
+}`
+	value := `{"name":"Peter Mueller"}`
+
+	topicName := test_util.CreateAvroTopic(t, "avro-topic", "", valueSchema)
+
+	kafkaCtl := test_util.CreateKafkaCtlCommand()
+
+	// produce valid avro message
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", value, "-H", "key1:value1", "-H", "key\\:2:value\\:2"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	test_util.AssertEquals(t, "message produced (partition=0\toffset=0)", kafkaCtl.GetStdOut())
+
+	// produce message that cannot be deserialized
+	test_util.SwitchContext("no-avro")
+
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", "no-avro"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	test_util.AssertEquals(t, "message produced (partition=0\toffset=1)", kafkaCtl.GetStdOut())
+
+	test_util.SwitchContext("default")
+
+	// produce another valid avro message
+	if _, err := kafkaCtl.Execute("produce", topicName, "--key", "test-key", "--value", value, "-H", "key1:value1", "-H", "key\\:2:value\\:2"); err != nil {
+		t.Fatalf("failed to execute command: %v", err)
+	}
+
+	test_util.AssertEquals(t, "message produced (partition=0\toffset=2)", kafkaCtl.GetStdOut())
+
+	if _, err := kafkaCtl.Execute("consume", topicName, "--from-beginning", "--exit"); err != nil {
+		test_util.AssertEquals(t, fmt.Sprintf("%s\n%s", value, value), kafkaCtl.GetStdOut())
+		test_util.AssertErrorContains(t, "failed to find avro schema", err)
+	} else {
+		t.Fatalf("expected consumer to fail")
+	}
 }