+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - SASL mechanism support now also implemented for cluster admin
+- process no longer gets stuck when deserialization error occurs 
 
 ## 1.16.0 - 2021-02-25